+++ sample/app/src/test/java/com/squareup/anvil/sample/UserComponentTest.kt
@@ -1,5 +1,6 @@
 package com.squareup.anvil.sample
 
+import com.google.common.truth.Truth.assertThat
 import org.junit.Test
 
 class UserComponentTest {
@@ -9,7 +10,7 @@ class UserComponentTest {
     val baseComponent = parent.user().create()
     val userComponent = baseComponent as UserComponent
 
-    assert(baseComponent.description() == UserDescriptionModule.provideDescription())
-    assert(userComponent.username() == UserDescriptionModule.provideName())
+    assertThat(baseComponent.description()).isEqualTo(UserDescriptionModule.provideDescription())
+    assertThat(userComponent.username()).isEqualTo(UserDescriptionModule.provideName())
   }
 }