+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 **Unreleased**
 --------------
 
-### Deprecated
-- `ClassReference.functions` has been deprecated in favor of `ClassReference.memberFunctions` and `ClassReference.declaredMemberFunctions`
+- **Fix**: Fix missing Jakarta `@Inject`-annotated constructors in factory gen.
+- **Deprecated**: `ClassReference.functions` has been deprecated in favor of `ClassReference.memberFunctions` and `ClassReference.declaredMemberFunctions`
 
 0.4.0
 -----